Move StreamTransformer factories
